requirements.txt
flask
pillow
requests
newspaper3k
openai
python-dotenv
lxml_html_clean